moved search bar style into flappy search bar file
