--- lib/search_bar_style.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SearchBarStyle {
-  final Color backgroundColor;
-  final EdgeInsetsGeometry padding;
-  final BorderRadius borderRadius;
-
-  const SearchBarStyle(
-      {this.backgroundColor = const Color.fromRGBO(142, 142, 147, .15),
-      this.padding = const EdgeInsets.all(5.0),
-      this.borderRadius: const BorderRadius.all(Radius.circular(5.0))});
-}
